Move TetherBase and use imports
There was some weirdness leftover from when TetherBase was global. This moves all the utils that were added to TetherBase.modules to using actual exports and building the modules directly.
