+++ src/js/utils/parents.js
@@ -1,11 +1,4 @@
-import { extend } from './utils/general';
-
-let TetherBase;
-if (typeof TetherBase === 'undefined') {
-  TetherBase = { modules: [] };
-}
-
-function getScrollParents(el) {
+export function getScrollParents(el) {
   // In firefox if the el is inside an iframe with display: none; window.getComputedStyle() will return null;
   // https://bugzilla.mozilla.org/show_bug.cgi?id=548397
   const computedStyle = getComputedStyle(el) || {};
@@ -48,56 +41,6 @@ function getScrollParents(el) {
   return parents;
 }
 
-function getOffsetParent(el) {
+export function getOffsetParent(el) {
   return el.offsetParent || document.documentElement;
 }
-
-let _scrollBarSize = null;
-
-function getScrollBarSize() {
-  if (_scrollBarSize) {
-    return _scrollBarSize;
-  }
-  const inner = document.createElement('div');
-  inner.style.width = '100%';
-  inner.style.height = '200px';
-
-  const outer = document.createElement('div');
-  extend(outer.style, {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    pointerEvents: 'none',
-    visibility: 'hidden',
-    width: '200px',
-    height: '150px',
-    overflow: 'hidden'
-  });
-
-  outer.appendChild(inner);
-
-  document.body.appendChild(outer);
-
-  const widthContained = inner.offsetWidth;
-  outer.style.overflow = 'scroll';
-  let widthScroll = inner.offsetWidth;
-
-  if (widthContained === widthScroll) {
-    widthScroll = outer.clientWidth;
-  }
-
-  document.body.removeChild(outer);
-
-  const width = widthContained - widthScroll;
-
-  _scrollBarSize = { width, height: width };
-  return _scrollBarSize;
-}
-
-TetherBase.Utils = {
-  getScrollParents,
-  getOffsetParent,
-  getScrollBarSize
-};
-
-export default TetherBase;
